paquo.objects.projects: factor out images
